requirements.txt
# 核心依赖
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
flask-migrate==4.1.0
flask-bootstrap==3.3.7.1
flask-ckeditor==1.0.0
flask-script==2.0.6

# 数据库
mysql-connector-python==9.2.0
sqlalchemy==2.0.40
alembic==1.15.2
greenlet==3.1.1

# 表单处理
wtforms==3.2.1
email-validator==2.2.0

# 文件处理
pillow==11.2.1
openpyxl==3.1.5
reportlab==4.4.0
et-xmlfile==2.0.0

# 模板与渲染
jinja2==3.1.6
mako==1.3.9
dominate==2.9.1
markupsafe==3.0.2

# 其他工具
python-dotenv==1.1.0
werkzeug==3.1.3
click==8.1.8
itsdangerous==2.2.0
blinker==1.9.0
visitor==0.1.3

# 网络与安全
chardet==5.2.0
dnspython==2.7.0
idna==3.10
typing-extensions==4.13.1
bcrypt==4.3.0